refactor(desktop): extract browser observability helpers
Move the pure observability layer — types + formatting / parsing helpers
that the per-tab listeners and HTTP routes both depend on — into a new
module browser-pane/observability.ts. Stateful per-tab append/upsert
helpers stay in main.ts (they need the tab record + workspace map).

Moved:
- types: BrowserConsoleLevel, BrowserErrorSource, BrowserConsoleEntry,
  BrowserObservedError, BrowserRequestBodyMetadata,
  BrowserResponseBodyMetadata, BrowserRequestRecord
- constants: BROWSER_OBSERVABILITY_ENTRY_LIMIT, BROWSER_REQUEST_HISTORY_LIMIT,
  BROWSER_OBSERVABILITY_DEFAULT_LIMIT
- pure fns: browserObservabilityLimit, browserConsoleLevelValue,
  browserConsoleLevelRank, browserObservedErrorSource,
  browserIsoFromNetworkTimestamp, browserHeaderRecord,
  browserHeaderFirstValue, browserResponseBodyMetadata,
  browserRequestBodyMetadata, appendBoundedEntry, browserRequestIdValue,
  browserRequestFailure, browserRequestSummary

main.ts -195. Typecheck clean.
